requirements.txt
asgiref==3.3.1
Django==3.1.3
django-colorfield==0.3.2
django-js-asset==1.2.2
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
